Update2 Denisa.md
Added a few more words about the mentorship and guidance we provide. Added a few more details about articles, how they are a starting point, and removed the [back to top] (otherwise coded for as: [Back to top](#index-content-by-topic) ) bc we don't even have that much text for them to be useful.